requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pdf2image>=1.16.0
pillow>=9.0.0
pyzbar>=0.1.9